% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicero_utils.R
\name{district_type_args_list}
\alias{district_type_args_list}
\title{district_type_args_list}
\usage{
district_type_args_list(x)
}
\arguments{
\item{x}{list from parameters}
}
\value{
list
}
\description{
format the arguments into acceptable query
}
\examples{
#' district_type_args_list(district_type)
}

