% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_cicero.R
\name{explore_cicero}
\alias{explore_cicero}
\title{Shiny app for exploring Cicero data on elected officials}
\usage{
explore_cicero()
}
\description{
Shiny app for exploring Cicero data on elected officials
}
\examples{

explore_cicero()

}

