% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legislative_district.R
\name{get_legislative_district}
\alias{get_legislative_district}
\title{get_legislative_district}
\usage{
get_legislative_district(search_loc = NULL, lat = NULL, lon = NULL)
}
\arguments{
\item{search_loc}{A string with most of any street address. Most of the time, you don't need to include zip code.}
}
\value{
data_frame
}
\description{
get_legislative_district
}
\examples{
#' get_legislative_district(search_loc = "3175 Bowers Ave. Santa Clara, CA")
#' get_legislative_district(lat = 40, lon = -75.1)
}

