% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_map}
\alias{get_map}
\title{get_map}
\usage{
get_map(state, country = "US", district_id, district_type)
}
\arguments{
\item{state}{string indicating state (e.g. "CA", "MA", "MZ", etc.)}

\item{country}{string indicating country (e.g. "MX", "CN", "UK")}

\item{district_id}{string indicating district id (obtainable through get_official(), get_legislative_district())}

\item{district_type}{string indicating district type (e.g. "NATIONAL_LOWER", "NATIONAL_UPPER", "STATE_LOWER", etc.)}
}
\value{
list
}
\description{
Obtain district map data by state, country, district type, and district id.
}
\examples{
#' map <- get_map(state = "CA", district_type = "NATIONAL_LOWER", district_id = 5)
}

