% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_upcoming_elections.R
\name{get_upcoming_elections}
\alias{get_upcoming_elections}
\title{get_upcoming_elections}
\usage{
get_upcoming_elections(election_expire_date_on_or_after = "today",
  is_local = NULL, is_state = NULL, is_national = NULL,
  is_transnational = NULL, elections = 5)
}
\arguments{
\item{election_expire_date_on_or_after}{a character vector indicating date. Defaults to "today"}

\item{is_local}{Boolean indicating that the ElectionEvent represents a local election}

\item{is_state}{Boolean indicating that the ElectionEvent represents a state election}

\item{is_national}{Boolean indicating that the ElectionEvent represents a national election}

\item{is_transnational}{Boolean indicating that the ElectionEvent represents a transnational election}

\item{elections}{an integer, indicating max number of elections to return}
}
\value{
data_frame
}
\description{
get_upcoming_elections
}
\examples{
#' get_upcoming_elections(is_state = TRUE, elections = 2)
#' get_upcoming_elections(is_national = TRUE, elections = 5)
}

