% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcicero-package.r
\docType{package}
\name{rcicero}
\alias{rcicero}
\alias{rcicero-package}
\title{rcicero: A Wrapper for the Cicero API}
\description{
This package makes getting data from the Cicero API easier for useRs. The objects returned by its functions are admittedly messy data frames, and some of the data may require unnesting, but it includes everything available from Cicero.
}

