% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicero_utils.R
\name{set_token_and_user}
\alias{set_token_and_user}
\title{set_token_and_user}
\usage{
set_token_and_user(username, password)
}
\arguments{
\item{username}{string - your Cicero account username}

\item{password}{string - your Cicero account password}
}
\value{
list
}
\description{
Obtain your Cicero API token and user with a POST request
}
\examples{
#' set_token_and_user("darazolin@gmail.com", "password")
}

