% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nonlegislative_district.R
\name{get_nonlegislative_district}
\alias{get_nonlegislative_district}
\title{get_nonlegislative_district}
\usage{
get_nonlegislative_district(address, type)
}
\arguments{
\item{address}{a string containing a street address}

\item{type}{a string, nonlegislative district type that is either: CENSUS, COUNTY, JUDICIAL, POLICE, SCHOOL, or WATERSHED}
}
\value{
data_frame
}
\description{
get_nonlegislative_district
}
\examples{
#' get_nonlegislative_district(address = "3175 Bowers Ave. Santa Clara, CA, type = "SCHOOL")
}

