% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_official.R
\name{get_official}
\alias{get_official}
\title{get_official}
\usage{
get_official(search_loc = NULL, lat = NULL, lon = NULL,
  first_name = NULL, last_name = NULL, valid_on = NULL,
  district_type = c("STATE_LOWER", "STATE_UPPER", "NATIONAL_UPPER",
  "NATIONAL_LOWER", "NATIONAL_EXEC"))
}
\arguments{
\item{search_loc}{Street address}

\item{lat}{Latitude coordiante}

\item{lon}{Longitude coordinate}

\item{first_name}{First name of official}

\item{last_name}{Last name of official}

\item{valid_on}{date for official}

\item{district_type}{Officials from which district type?}
}
\value{
list
}
\description{
Obtain official information from the Cicero API. Can seach by latitude and longitude coordinates and first and last name.
}
\examples{
#' get_official(lat = 40, lon = -75.1)
#' get_official(search_loc = "3175 Bowers Ave. Santa Clara, CA", district_type = "STATE_LOWER", "STATE_UPPER")
#' get_official(last_name = "Obama")
}

